pyproject.toml
[project]
name = "aimaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
